Fixed a bug in changeLogResult
checkout calls changeLogResult with a null-changelog, but changeLogResult handles this as an error and returns false. This false is delivered directly to Hudson which is aborting the build if it was started without any commit. We had to ways to fix this: Let checkout always return true on now changes or change changeLogResult to always return true. I preferred the latter because the intended call of changeLogResult seems like deleting the changeLogFile.